feature: added new fields for user refactor: updated start screen and auth ones refactor: refactored button, input and title components

+++ lib/src/models/profiles/user_profile.dart
@@ -31,7 +31,7 @@ class UserProfile extends Profile {
   late String gender;
   late String? education;
   late String? favoriteSong;
-  late double? height;
+  late int? height;
   late List<SocialNetwork>? socialNetworks;
 
   UserProfile({
