+++ lib/src/screens/user_profile_detailed_view/character_trait_view.dart
@@ -26,7 +26,7 @@ class CharacterTraitView extends StatelessWidget {
           ],
         ),
         Slider(
-          value: characterTrait.degree,
+          value: characterTrait.degree.toDouble(),
           min: 0,
           max: 10,
           activeColor: Theme.of(context).colorScheme.onBackground,
